+++ android_lib/src/main/java/io/sentry/godotplugin/SentryAndroidGodotPlugin.kt
@@ -6,10 +6,15 @@ import io.sentry.Breadcrumb
 import io.sentry.Hint
 import io.sentry.ISerializer
 import io.sentry.Sentry
+import io.sentry.SentryAttributes
 import io.sentry.SentryEvent
 import io.sentry.SentryLevel
+import io.sentry.SentryLogEvent
+import io.sentry.SentryLogEventAttributeValue
+import io.sentry.SentryLogLevel
 import io.sentry.SentryOptions
 import io.sentry.android.core.SentryAndroid
+import io.sentry.logger.SentryLogParameters
 import io.sentry.protocol.Feedback
 import io.sentry.protocol.Message
 import io.sentry.protocol.SentryException
@@ -51,6 +56,12 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
         }
     }
 
+    private val logsByHandle = object : ThreadLocal<MutableMap<Int, SentryLogEvent>>() {
+        override fun initialValue(): MutableMap<Int, SentryLogEvent> {
+            return mutableMapOf()
+        }
+    }
+
     private fun getEvent(eventHandle: Int): SentryEvent? {
         val event: SentryEvent? = eventsByHandle.get()?.get(eventHandle)
         if (event == null) {
@@ -73,6 +84,14 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
             Log.e(TAG, "Internal Error -- Breadcrumb not found: $breadcrumbHandle")
         }
         return crumb
+    }
+
+    private fun getLog(logHandle: Int): SentryLogEvent? {
+        var logEvent: SentryLogEvent? = logsByHandle.get()?.get(logHandle)
+        if (logEvent == null) {
+            Log.e(TAG, "Internal Error -- SentryLogEvent not found: $logHandle")
+        }
+        return logEvent
     }
 
     private fun registerEvent(event: SentryEvent): Int {
@@ -120,6 +139,21 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
         return handle
     }
 
+    private fun registerLog(logEvent: SentryLogEvent): Int {
+        var logsMap = logsByHandle.get() ?: run {
+            Log.e(TAG, "Internal Error -- logsMap is null")
+            return 0
+        }
+
+        var handle = Random.nextInt()
+        while (logsMap.containsKey(handle)) {
+            handle = Random.nextInt()
+        }
+
+        logsMap[handle] = logEvent
+        return handle
+    }
+
     override fun getPluginName(): String {
         return "SentryAndroidGodotPlugin"
     }
@@ -134,6 +168,8 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
         environment: String,
         sampleRate: Float,
         maxBreadcrumbs: Int,
+        enableLogs: Boolean,
+        beforeSendLogHandlerId: Long
     ) {
         Log.v(TAG, "Initializing Sentry Android")
         SentryAndroid.init(godot.getActivity()!!.applicationContext) { options ->
@@ -146,6 +182,7 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
             options.maxBreadcrumbs = maxBreadcrumbs
             options.sdkVersion?.name = "sentry.java.android.godot"
             options.nativeSdkName = "sentry.native.android.godot"
+            options.logs.isEnabled = enableLogs
             options.beforeSend =
                 SentryOptions.BeforeSendCallback { event: SentryEvent, hint: Hint ->
                     Log.v(TAG, "beforeSend: ${event.eventId} isCrashed: ${event.isCrashed}")
@@ -153,7 +190,16 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
                     Callable.call(beforeSendHandlerId, "before_send", handle)
                     eventsByHandle.get()?.remove(handle) // Returns the event or null if it was discarded.
                 }
+            if (beforeSendLogHandlerId != 0L) {
+                options.logs.beforeSend =
+                    SentryOptions.Logs.BeforeSendLogCallback { logEvent ->
+                        val handle: Int = registerLog(logEvent)
+                        Callable.call(beforeSendLogHandlerId, "before_send_log", handle)
+                        logsByHandle.get()?.remove(handle) // Returns the log or null if it was discarded.
+                    }
             }
+
+        }
     }
 
     @UsedByGodot
@@ -237,6 +283,20 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
     fun addBreadcrumb(handle: Int) {
         val crumb = getBreadcrumb(handle) ?: return
         Sentry.addBreadcrumb(crumb)
+    }
+
+    @UsedByGodot
+    fun log(level: Int, body: String, attributes: Dictionary) {
+        if (attributes.isEmpty()) {
+            Sentry.logger().log(level.toSentryLogLevel(), body)
+        } else {
+            val sentryAttributes = SentryAttributes.fromMap(attributes)
+            Sentry.logger().log(
+                level.toSentryLogLevel(),
+                SentryLogParameters.create(sentryAttributes),
+                body
+            )
+        }
     }
 
     @UsedByGodot
@@ -571,6 +631,74 @@ class SentryAndroidGodotPlugin(godot: Godot) : GodotPlugin(godot) {
     fun breadcrumbGetTimestamp(handle: Int): Long {
         val crumb = getBreadcrumb(handle) ?: return 0
         return crumb.timestamp.toMicros()
+    }
+
+    @UsedByGodot
+    fun releaseLog(handle: Int) {
+        val logsMap = logsByHandle.get() ?: run {
+            Log.e(TAG, "Internal Error -- logsByHandle is null")
+            return
+        }
+
+        logsMap.remove(handle)
+    }
+
+    @UsedByGodot
+    fun logSetLevel(handle: Int, level: Int) {
+        getLog(handle)?.level = level.toSentryLogLevel()
+    }
+
+    @UsedByGodot
+    fun logGetLevel(handle: Int): Int {
+        return getLog(handle)?.level?.toInt() ?: SentryLogLevel.INFO.toInt()
+    }
+
+    @UsedByGodot
+    fun logSetBody(handle: Int, body: String) {
+        getLog(handle)?.body = body
+    }
+
+    @UsedByGodot
+    fun logGetBody(handle: Int): String {
+        return getLog(handle)?.body ?: ""
+    }
+
+    @UsedByGodot
+    fun logGetAttribute(handle: Int, name: String): Dictionary {
+        // NOTE: Use Dictionary container for the value to avoid object wrapper creation.
+        var attr = getLog(handle)?.attributes?.get(name) ?: return Dictionary()
+        val result = Dictionary()
+        result["type"] = attr.type
+        result["value"] = attr.value
+        return result
+    }
+
+    @UsedByGodot
+    fun logSetAttribute(handle: Int, name: String, type: String, value: Any) {
+        val log = getLog(handle) ?: return
+        val logAttributes = log.attributes ?: HashMap<String, SentryLogEventAttributeValue>().also { log.attributes = it }
+        logAttributes[name] = SentryLogEventAttributeValue(type, value)
+    }
+
+    @UsedByGodot
+    fun logAddAttributes(handle: Int, attributes: Dictionary) {
+        val log = getLog(handle) ?: return
+        val logAttributes = log.attributes ?: HashMap<String, SentryLogEventAttributeValue>().also { log.attributes = it }
+        for ((key, value) in attributes) {
+            val attrValue = when(value) {
+                is Boolean -> SentryLogEventAttributeValue("boolean", value)
+                is Int, is Long -> SentryLogEventAttributeValue("integer", value)
+                is Float, is Double -> SentryLogEventAttributeValue("double", value)
+                is String -> SentryLogEventAttributeValue("string", value)
+                else -> SentryLogEventAttributeValue("string", value.toString())
+            }
+            logAttributes[key.toString()] = attrValue
+        }
+    }
+
+    @UsedByGodot
+    fun logRemoveAttribute(handle: Int, name: String) {
+        getLog(handle)?.attributes?.remove(name)
     }
 
 }