+++ android_lib/src/main/java/io/sentry/godotplugin/UtilityFunctions.kt
@@ -1,6 +1,7 @@
 package io.sentry.godotplugin
 
 import io.sentry.SentryLevel
+import io.sentry.SentryLogLevel
 import java.util.Date
 import java.time.Instant
 
@@ -14,6 +15,17 @@ fun Int.toSentryLevel(): SentryLevel =
         else -> SentryLevel.ERROR
     }
 
+fun Int.toSentryLogLevel(): SentryLogLevel =
+    when (this) {
+        0 -> SentryLogLevel.TRACE
+        1 -> SentryLogLevel.DEBUG
+        2 -> SentryLogLevel.INFO
+        3 -> SentryLogLevel.WARN
+        4 -> SentryLogLevel.ERROR
+        5 -> SentryLogLevel.FATAL
+        else -> SentryLogLevel.INFO
+    }
+
 fun SentryLevel.toInt(): Int =
     when (this) {
         SentryLevel.DEBUG -> 0
@@ -21,6 +33,16 @@ fun SentryLevel.toInt(): Int =
         SentryLevel.WARNING -> 2
         SentryLevel.ERROR -> 3
         SentryLevel.FATAL -> 4
+    }
+
+fun SentryLogLevel.toInt(): Int =
+    when (this) {
+        SentryLogLevel.TRACE -> 0
+        SentryLogLevel.DEBUG -> 1
+        SentryLogLevel.INFO -> 2
+        SentryLogLevel.WARN -> 3
+        SentryLogLevel.ERROR -> 4
+        SentryLogLevel.FATAL -> 5
     }
 
 fun Long.microsecondsToTimestamp(): Date {