+++ src/sentry/android/android_sdk.h
@@ -22,11 +22,27 @@ protected:
 	static void _bind_methods();
 };
 
+class SentryAndroidBeforeSendLogHandler : public Object {
+	GDCLASS(SentryAndroidBeforeSendLogHandler, Object);
+	friend class AndroidSDK;
+
+private:
+	Object *android_plugin = nullptr;
+
+	void _initialize(Object *p_android_plugin);
+
+	void _before_send_log(int32_t p_log_handle);
+
+protected:
+	static void _bind_methods();
+};
+
 // Internal SDK utilizing Sentry Android (sentry-java repo).
 class AndroidSDK : public InternalSDK {
 private:
 	Object *android_plugin = nullptr;
 	SentryAndroidBeforeSendHandler *before_send_handler = nullptr;
+	SentryAndroidBeforeSendLogHandler *before_send_log_handler = nullptr;
 
 public:
 	virtual void set_context(const String &p_key, const Dictionary &p_value) override;
@@ -40,6 +56,8 @@ public:
 
 	virtual Ref<SentryBreadcrumb> create_breadcrumb() override;
 	virtual void add_breadcrumb(const Ref<SentryBreadcrumb> &p_breadcrumb) override;
+
+	virtual void log(LogLevel p_level, const String &p_body, const Dictionary &p_attributes = Dictionary()) override;
 
 	virtual String capture_message(const String &p_message, Level p_level = sentry::LEVEL_INFO) override;
 	virtual String get_last_event_id() override;